moved things to folders, fixed part of checkbox functionality (habits stay checked now between restarts of the app)
