+++ app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Habbit">
         <activity
-            android:name=".HabitEventsActivity"
+            android:name=".activities.HabitEventsActivity"
             android:exported="true" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
