+++ app/src/main/java/com/example/habbit/fragments/HabitEntryFragment.java
@@ -1,9 +1,10 @@
-package com.example.habbit;
+package com.example.habbit.fragments;
 
 import android.app.AlertDialog;
 import android.app.DatePickerDialog;
 import android.app.Dialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
@@ -16,6 +17,9 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import com.example.habbit.models.Habit;
+import com.example.habbit.R;
 
 import java.io.Serializable;
 import java.text.SimpleDateFormat;
