+++ app/src/main/java/com/example/habbit/models/Habit.java
@@ -1,9 +1,7 @@
-package com.example.habbit;
+package com.example.habbit.models;
 
 import java.io.Serializable;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 
 public class Habit implements Serializable {
     /* attributes that define a Habit object */
@@ -15,7 +13,7 @@ public class Habit implements Serializable {
     /* to be implemented */
     // TODO: Should we use this arraylist, or just use Firestore to track habitEvents?
     private ArrayList<HabitEvent> habitEvents;
-    private boolean checked = false; // TODO: this is temporary, habit behaviour is too complicated for to think about right now
+    private boolean checked = false;
 
     // empty constructor needed to use Firestore add()
     public Habit() {}
