+++ app/src/main/java/com/example/habbit/models/User.java
@@ -1,4 +1,4 @@
-package com.example.habbit;
+package com.example.habbit.models;
 
 import java.util.ArrayList;
