+++ app/src/main/res/layout/activity_habit_events.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".HabitEventsActivity">
+    tools:context=".activities.HabitEventsActivity">
 
     <ListView
         android:id="@+id/habit_events_list"
