+++ app/src/main/res/layout/fragment_habit_entry.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HabitEntryFragment"
+    tools:context=".fragments.HabitEntryFragment"
     android:orientation="vertical"
     android:paddingHorizontal="30dp"
     android:paddingVertical="30dp">
