+++ app/src/main/res/layout/fragment_view_habit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HabitDetailsFragment"
+    tools:context=".fragments.HabitDetailsFragment"
     android:orientation="vertical"
     android:paddingHorizontal="30dp"
     android:paddingVertical="30dp">
